python_pdf_service/requirements.txt
# PDF Processing Microservice Dependencies
# Install with: pip install -r requirements.txt

# Core PDF processing
PyMuPDF>=1.26.0

# Image processing
Pillow>=10.0.0

# Web framework
Flask>=3.0.0

# CORS support
flask-cors>=4.0.0

# HTTP requests (for future integrations)
requests>=2.31.0

